+++ src/test/java/dandi/dandi/notification/application/service/eventhandler/PostLikeNotificationEventHandlerTest.java
@@ -1,4 +1,4 @@
-package dandi.dandi.notification.application;
+package dandi.dandi.notification.application.service.eventhandler;
 
 import static dandi.dandi.member.MemberTestFixture.MEMBER_ID;
 import static dandi.dandi.post.PostFixture.POST_ID;
